[Gold V] Title: 1, 2, 3 더하기 4, Time: 0 ms, Memory: 2180 KB -BaekjoonHub

+++ 백준/Gold/15989. 1， 2， 3 더하기 4/1， 2， 3 더하기 4.cc
@@ -7,7 +7,7 @@ using namespace std;
 int main(){
     FASTIO
 
-    vector<long long> dp(10'001, 0ll);
+    vector<int> dp(10'001, 0);
     dp[0] = 1;
 
     for(int i = 1; i <= 3; i++){
